Use package name as basename for Thor classes
