+++ lib/cpflow.rb
@@ -31,25 +31,7 @@ at_exit do
   end
 end
 
-# Fix for https://github.com/erikhuda/thor/issues/398
-# Copied from https://github.com/rails/thor/issues/398#issuecomment-622988390
-class Thor
-  module Shell
-    class Basic
-      def print_wrapped(message, options = {})
-        indent = (options[:indent] || 0).to_i
-        if indent.zero?
-          stdout.puts(message)
-        else
-          message.each_line do |message_line|
-            stdout.print(" " * indent)
-            stdout.puts(message_line.chomp)
-          end
-        end
-      end
-    end
-  end
-end
+require_relative "patches/thor"
 
 module Cpflow
   class Error < StandardError; end
